+++ composeApp/src/commonMain/kotlin/com/example/bookpedia/book/data/mappers/BookMappers.kt
@@ -1,5 +1,6 @@
 package com.example.bookpedia.book.data.mappers
 
+import com.example.bookpedia.book.data.database.BookEntity
 import com.example.bookpedia.book.data.dto.SearchBookDto
 import com.example.bookpedia.book.domain.Book
 
@@ -17,5 +18,37 @@ fun SearchBookDto.toBook(): Book {
     ratingCount = ratingsCount,
     numPages = numPagesMedian,
     numEditions = numEditions ?: 0
+  )
+}
+
+fun Book.toBookEntity(): BookEntity {
+  return BookEntity(
+    id = id,
+    title = title,
+    description = description,
+    imageUrl = imageUrl,
+    languages = languages,
+    authors = authors,
+    firstPublishYear = firstPublishYear,
+    ratingsAverage = averageRating,
+    ratingsCount = ratingCount,
+    numPagesMedian = numPages,
+    numEditions = numEditions
+  )
+}
+
+fun BookEntity.toBook(): Book {
+  return Book(
+    id = id,
+    title = title,
+    description = description,
+    imageUrl = imageUrl,
+    languages = languages,
+    authors = authors,
+    firstPublishYear = firstPublishYear,
+    averageRating = ratingsAverage,
+    ratingCount = ratingsCount,
+    numPages = numPagesMedian,
+    numEditions = numEditions
   )
 }