+++ composeApp/src/commonMain/kotlin/com/example/bookpedia/book/data/repository/DefaultBookRepository.kt
@@ -1,15 +1,22 @@
 package com.example.bookpedia.book.data.repository
 
+import androidx.sqlite.SQLiteException
+import com.example.bookpedia.book.data.database.FavoriteBooksDao
 import com.example.bookpedia.book.data.mappers.toBook
+import com.example.bookpedia.book.data.mappers.toBookEntity
 import com.example.bookpedia.book.data.network.RemoteBookDataSource
 import com.example.bookpedia.book.domain.Book
 import com.example.bookpedia.book.domain.BookRepository
 import com.example.bookpedia.core.domain.DataError
+import com.example.bookpedia.core.domain.EmptyResult
 import com.example.bookpedia.core.domain.Result
 import com.example.bookpedia.core.domain.map
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
 
 class DefaultBookRepository(
-  private val remoteBookDataSource: RemoteBookDataSource
+  private val remoteBookDataSource: RemoteBookDataSource,
+  private val favoriteBooksDao: FavoriteBooksDao
 ) : BookRepository {
   override suspend fun searchBooks(query: String): Result<List<Book>, DataError.Remote> {
     return remoteBookDataSource
@@ -19,9 +26,42 @@ class DefaultBookRepository(
       }
   }
 
-  override suspend fun getBookDescription(bookWorkId: String): Result<String?, DataError> {
-    return remoteBookDataSource
-      .getBookDetails(bookWorkId)
-      .map { it.description }
+  override suspend fun getBookDescription(bookId: String): Result<String?, DataError> {
+    val localResult = favoriteBooksDao.getFavoriteBook(bookId)
+
+     return if(localResult == null) {
+       remoteBookDataSource
+        .getBookDetails(bookId)
+        .map { it.description }
+    } else Result.Success(localResult.description)
+  }
+
+  override fun getFavoriteBooks(): Flow<List<Book>> {
+    return favoriteBooksDao
+      .getFavoriteBooks()
+      .map { bookEntities ->
+        bookEntities.map { it.toBook() }
+      }
+  }
+
+  override fun isBookFavorite(id: String): Flow<Boolean> {
+    return favoriteBooksDao
+      .getFavoriteBooks()
+      .map { bookEntities ->
+        bookEntities.any { it.id == id }
+      }
+  }
+
+  override suspend fun markAsFavorite(book: Book): EmptyResult<DataError.Local> {
+    return try {
+      favoriteBooksDao.upsert(book.toBookEntity())
+      Result.Success(Unit)
+    } catch (_: SQLiteException) {
+      Result.Error(DataError.Local.DISK_FULL)
+    }
+  }
+
+  override suspend fun unmarkAsFavorite(id: String) {
+    favoriteBooksDao.deleteFavoriteBook(id)
   }
 }