+++ composeApp/src/commonMain/kotlin/com/example/bookpedia/book/presentation/book_details/BookDetailsScreen.kt
@@ -1,7 +1,6 @@
 package com.example.bookpedia.book.presentation.book_details
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row