+++ composeApp/src/commonMain/kotlin/com/example/bookpedia/book/presentation/book_list/BookListViewModel.kt
@@ -27,6 +27,7 @@ class BookListViewModel(
 
   private var cachedBooks = emptyList<Book>()
   private var searchJob: Job? = null
+  private var observeFavoriteBooksJob: Job? = null
 
   private val _state = MutableStateFlow(BookListState())
   val state = _state
@@ -34,6 +35,8 @@ class BookListViewModel(
       if (cachedBooks.isEmpty()) {
         observeSearchQuery()
       }
+
+      observeFavoriteBooks()
     }
     .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000L), _state.value)
 
@@ -54,6 +57,16 @@ class BookListViewModel(
         }
       }
     }
+  }
+
+  private fun observeFavoriteBooks() {
+    observeFavoriteBooksJob?.cancel()
+    observeFavoriteBooksJob = bookRepository
+      .getFavoriteBooks()
+      .onEach { favoriteBooks -> _state.update {
+          it.copy(favoriteBooks = favoriteBooks)
+        } }
+      .launchIn(viewModelScope)
   }
 
   @OptIn(FlowPreview::class)